+++ ImporterContext.hpp
@@ -117,8 +117,8 @@ class ImporterContext final : public IImporterContext
     //! Map holding FunctionProtos
     StringMap<::ONNX_NAMESPACE::FunctionProto> mLocalFunctions;
 
-    //! Vector to hold current local function names
-    std::vector<std::string> mLocalFunctionStack;
+    //! Vector to hold current local function names and attributes
+    std::vector<std::pair<std::string, StringMap<::ONNX_NAMESPACE::AttributeProto const*>>> mLocalFunctionStack;
 
     //! Vector to hold expected graph outputs
     std::vector<::ONNX_NAMESPACE::ValueInfoProto> mGraphOutputNames;
@@ -336,7 +336,7 @@ public:
     {
         return mLocalFunctions;
     }
-    std::vector<std::string>& localFunctionStack() override
+    std::vector<std::pair<std::string, StringMap<::ONNX_NAMESPACE::AttributeProto const*>>>& localFunctionStack() override
     {
         return mLocalFunctionStack;
     }